Tag collection indices are no longer stored in the session, all tag collections are randomized.
Tag collections are randomized by default now.
Removed canUpdateState: it wasn't good enough.
Added a backup of the tag collection indices. This is still not good enough.
Always telling Flask the session was modified, to be sure.
Completely removed storing of lists of indices in the session: it's too risky.